IsThereAnyDeal game-specific links on EpicGames Store v0.9.2
Stopped clearing ButtonStayInterval because sometimes when loading a store page, the ITAD button would get inserted before the page fully loads (used window.onload, but it's not good enough due to dynamic content) and would then get removed by the time the page is fully loaded. So now it works how it used to work, incidentally matching the function name "InsertAndMakeSureButtonStays".
This could/should be later updated "properly" by detecting when the page is actually ready for the button to be added.

Also, changed some let to const.