IMDb Youtube Trailer Button v0.3.3: year issue
Fixed issue where year selector would not find the year properly with titles that had an alternative title.
Should now find it in those cases as well.
Also, if for some reason the year is still not found, the parentheses (or year obviously) are not added to the search string.